Unmodified JoobyPlugin, but moved to org.mytake.gradle.jooby to make sure that our modifications run as expected.
