+++ buildSrc/src/main/java/org/mytake/gradle/jooby/AssetTask.java
@@ -254,13 +254,14 @@
  * containing Woodstox, in file "ASL2.0", under the same directory
  * as this file.
  */
-package org.jooby.run;
+package org.mytake.gradle.jooby;
 
 import java.io.File;
 import java.io.FileWriter;
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
 import java.util.Map;
+import java.util.function.Consumer;
 import java.util.stream.Collectors;
 
 import org.gradle.api.internal.ConventionTask;
@@ -271,6 +272,7 @@ import org.gradle.api.tasks.OutputFile;
 import org.gradle.api.tasks.TaskAction;
 import org.jooby.assets.AssetCompiler;
 
+import com.google.common.base.Throwables;
 import com.google.common.io.Files;
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
