+++ buildSrc/src/main/java/org/mytake/gradle/jooby/JoobyProject.java
@@ -201,154 +201,119 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package org.jooby.run;
+/**
+ * This copy of Woodstox XML processor is licensed under the
+ * Apache (Software) License, version 2.0 ("the License").
+ * See the License for details about distribution rights, and the
+ * specific rights regarding derivate works.
+ *
+ * You may obtain a copy of the License at:
+ *
+ * http://www.apache.org/licenses/
+ *
+ * A copy is also included in the downloadable source code package
+ * containing Woodstox, in file "ASL2.0", under the same directory
+ * as this file.
+ */
+package org.mytake.gradle.jooby;
 
-import org.gradle.api.Plugin;
 import org.gradle.api.Project;
-import org.gradle.api.Task;
-import org.gradle.api.internal.ConventionMapping;
-import org.gradle.api.invocation.Gradle;
+import org.gradle.api.artifacts.Configuration;
+import org.gradle.api.plugins.JavaPluginConvention;
+import org.gradle.api.tasks.SourceSet;
+import static org.jooby.funzy.Throwing.throwingFunction;
 
 import java.io.File;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.Map;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.net.URLClassLoader;
+import java.util.LinkedHashSet;
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 
-public class JoobyPlugin implements Plugin<Project> {
+public class JoobyProject {
 
-  static {
-    // Gradle hack: OS j2v8 dependency
-    String family = os(System.getProperty("os.name", "").toLowerCase());
-    String arch = osarch(System.getProperty("os.arch", "").toLowerCase());
-    Object j2v8 = "j2v8_" + family + "_" + arch;
-    System.getProperties().put("j2v8", j2v8);
+  private Project project;
+
+  public JoobyProject(final Project project) {
+    this.project = project;
   }
 
-  @Override
-  public void apply(final Project project) {
-    configureJoobyRun(project);
-
-    configureJoobyAssets(project);
-
-    configureJoobySpec(project);
-
-    configureApiTool(project);
+  /**
+   * @return Returns the output directory for .class file.
+   */
+  public File classes() {
+    SourceSet sourceSet = sourceSet(project);
+    return sourceSet.getRuntimeClasspath().getFiles().stream()
+        .filter(f -> f.exists() && f.isDirectory() && f.toString().contains("classes"))
+        .findFirst()
+        .get();
   }
 
-  private void configureJoobyRun(final Project project) {
-    project.getTasks()
-        .withType(JoobyTask.class, joobyRun -> {
-          ConventionMapping mapping = joobyRun.getConventionMapping();
+  public Set<File> classpath() {
+    SourceSet sourceSet = sourceSet(project);
 
-          mapping.map("classpath", () -> new JoobyProject(project).classpath());
+    Set<File> cp = new LinkedHashSet<>();
+    // conf & public
+    sourceSet.getResources().getSrcDirs().forEach(cp::add);
 
-          mapping.map("src", () -> new JoobyProject(project).sources());
+    // classes/main, resources/main + jars
+    sourceSet.getRuntimeClasspath().getFiles().forEach(cp::add);
 
-          mapping.map("mainClassName", () -> project.getProperties().get("mainClassName"));
-
-          mapping.map("srcExtensions", () -> Arrays.asList(".java", ".conf", ".properties", ".kt"));
-
-          mapping.map("compiler", () -> {
-            File eclipseClasspath = new File(project.getProjectDir(), ".classpath");
-            return eclipseClasspath.exists() ? "off" : "on";
-          });
-
-          Gradle gradle = project.getGradle();
-          mapping.map("block", () -> !gradle.getStartParameter().isContinuous());
-          mapping.map("logLevel", () -> gradle.getStartParameter().getLogLevel().name());
-        });
-
-    Map<String, Object> options = new HashMap<>();
-    options.put(Task.TASK_TYPE, JoobyTask.class);
-    options.put(Task.TASK_DEPENDS_ON, "classes");
-    options.put(Task.TASK_NAME, "joobyRun");
-    options.put(Task.TASK_DESCRIPTION, "Run, debug and hot reload applications");
-    options.put(Task.TASK_GROUP, "jooby");
-    project.getTasks().create(options);
-  }
-
-  private void configureJoobyAssets(final Project project) {
-    project.getTasks()
-        .withType(AssetTask.class, task -> {
-          ConventionMapping mapping = task.getConventionMapping();
-
-          mapping.map("env", () -> "dist");
-
-          mapping.map("maxAge", () -> "365d");
-
-          mapping.map("mainClassName", () -> project.getProperties().get("mainClassName"));
-
-          mapping.map("output", () -> new JoobyProject(project).classes());
-
-          mapping.map("assemblyOutput", () -> new File(project.getBuildDir(), "__public_"));
-        });
-
-    Map<String, Object> options = new HashMap<>();
-    options.put(Task.TASK_TYPE, AssetTask.class);
-    options.put(Task.TASK_DEPENDS_ON, "classes");
-    options.put(Task.TASK_NAME, "joobyAssets");
-    options.put(Task.TASK_DESCRIPTION, "Process, optimize and compress static files");
-    options.put(Task.TASK_GROUP, "jooby");
-    project.getTasks().create(options);
-  }
-
-  private void configureJoobySpec(final Project project) {
-    project.getTasks()
-        .withType(SpecTask.class, task -> {
-          ConventionMapping mapping = task.getConventionMapping();
-
-          mapping.map("mainClassName", () -> project.getProperties().get("mainClassName"));
-
-          mapping.map("source", () -> new JoobyProject(project).javaSrc());
-        });
-
-    Map<String, Object> options = new HashMap<>();
-    options.put(Task.TASK_TYPE, SpecTask.class);
-    options.put(Task.TASK_DEPENDS_ON, "classes");
-    options.put(Task.TASK_NAME, "joobySpec");
-    options.put(Task.TASK_DESCRIPTION, "Export your API/microservices outside a Jooby application");
-    options.put(Task.TASK_GROUP, "jooby");
-    project.getTasks().create(options);
-  }
-
-  private void configureApiTool(final Project project) {
-    project.getTasks()
-        .withType(ApiToolTask.class, task -> {
-          ConventionMapping mapping = task.getConventionMapping();
-          mapping.map("mainClassName", () -> project.getProperties().get("mainClassName"));
-        });
-
-    Map<String, Object> options = new HashMap<>();
-    options.put(Task.TASK_TYPE, ApiToolTask.class);
-    options.put(Task.TASK_DEPENDS_ON, "classes");
-    options.put(Task.TASK_NAME, "joobyApiTool");
-    options
-        .put(Task.TASK_DESCRIPTION, "Export your HTTP API to open standards like Swagger and RAML");
-    options.put(Task.TASK_GROUP, "jooby");
-
-    project.getTasks().create(options);
-  }
-
-  private static String os(final String os) {
-    if (os.contains("windows")) {
-      return "win32";
-    } else if (os.contains("linux")) {
-
-      return "linux";
-    } else if (os.contains("mac")) {
-      return "macosx";
+    // provided?
+    Configuration provided = project.getConfigurations().findByName("provided");
+    if (provided != null) {
+      provided.getFiles().forEach(cp::add);
     }
-    return os;
+
+    return cp;
   }
 
-  private static String osarch(final String arch) {
-    if (arch.contains("x86_64")) {
-      return "x86_64";
-    } else if (arch.contains("x86")) {
-      return "x86";
-    } else if (arch.contains("amd64")) {
-      return "amd64";
-    }
-    return arch;
+  public Set<File> sources() {
+    SourceSet sourceSet = sourceSet(project);
+
+    Set<File> src = new LinkedHashSet<>();
+    // conf & public
+    sourceSet.getResources().getSrcDirs().forEach(src::add);
+
+    // source java: always add parent file: should be src/main
+    sourceSet.getJava().getSrcDirs().forEach(f -> src.add(f.getParentFile()));
+
+    return src;
+  }
+
+  public File javaSrc() {
+    SourceSet sourceSet = sourceSet(project);
+
+    // source java
+    return sourceSet.getJava().getSrcDirs().iterator().next();
+  }
+
+  private SourceSet sourceSet(final Project project) {
+    SourceSet sourceSet = getJavaConvention(project).getSourceSets()
+        .getByName(SourceSet.MAIN_SOURCE_SET_NAME);
+    return sourceSet;
+  }
+
+  public JavaPluginConvention getJavaConvention(final Project project) {
+    return project.getConvention().getPlugin(JavaPluginConvention.class);
+  }
+
+  public URLClassLoader newClassLoader() throws MalformedURLException {
+    return toClassLoader(
+        classpath().stream()
+            .map(throwingFunction(f -> f.toURI().toURL()))
+            .collect(Collectors.toList()),
+        getClass().getClassLoader());
+  }
+
+  private static URLClassLoader toClassLoader(final List<URL> cp, final ClassLoader parent) {
+    return new URLClassLoader(cp.toArray(new URL[cp.size()]), parent) {
+      @Override
+      public String toString() {
+        return cp.toString();
+      }
+    };
   }
 }
